+++ src/Exomia.Framework/ContentSerialization/ContentSerializer.cs
@@ -16,10 +16,10 @@ using Exomia.Framework.ContentSerialization.Exceptions;
 using Exomia.Framework.ContentSerialization.Readers;
 using Exomia.Framework.ContentSerialization.Types;
 using Exomia.Framework.ContentSerialization.Writers;
+using SharpDX;
 #if NETSTANDARD2_0
 using Exomia.Framework.Extensions;
 #endif
-using SharpDX;
 
 namespace Exomia.Framework.ContentSerialization
 {
