+++ src/Exomia.Framework/Extensions/DeconstructExtensions.cs
@@ -1,4 +1,14 @@
-﻿#if NETSTANDARD2_0
+﻿#region License
+
+// Copyright (c) 2018-2020, exomia
+// All rights reserved.
+// 
+// This source code is licensed under the BSD-style license found in the
+// LICENSE file in the root directory of this source tree.
+
+#endregion
+
+#if NETSTANDARD2_0
 using System.Collections.Generic;
 
 namespace Exomia.Framework.Extensions
