+++ src/Exomia.Framework/Graphics/Model/Obj.cs
@@ -59,6 +59,11 @@ namespace Exomia.Framework.Graphics.Model
             Faces    = faces;
         }
 
+        internal static Obj Create(Vertex[] vertices, Normal[] normals, Texture[] textures, Face[] faces)
+        {
+            return new Obj(vertices, normals, textures, faces);
+        }
+
         /// <summary>
         ///     A vertex.
         /// </summary>
@@ -269,11 +274,6 @@ namespace Exomia.Framework.Graphics.Model
             {
                 return $"{string.Join(", ", Vertices)} [{Vertices.Length}]";
             }
-        }
-
-        internal static Obj Create(Vertex[] vertices, Normal[] normals, Texture[] textures, Face[] faces)
-        {
-            return new Obj(vertices, normals, textures, faces);
         }
     }
 }
