+++ src/Exomia.Framework/Graphics/Shader/ShaderFileLoader.cs
@@ -37,6 +37,102 @@ namespace Exomia.Framework.Graphics.Shader
             "^\\s*\\*\\s*(vs|ps|ds|gs|hs|cs)\\s*([^\\s]+)\\s*([^\\s]+)\\s(.*)$",
             RegexOptions.Compiled | RegexOptions.Singleline);
 
+        public static Shader? FromStream(IGraphicsDevice graphicsDevice, Stream stream)
+        {
+            using StreamReader sr = new StreamReader(stream, Encoding.Default, true, 4096, true);
+            if (!SHADER_DEFINITION.Equals(sr.ReadLine()?.Trim(), StringComparison.InvariantCultureIgnoreCase))
+            {
+                return null;
+            }
+
+            IList<Group> groups = new List<Group>(1);
+
+            Group?  currentGroup = null;
+            string? line;
+            while ((line = sr.ReadLine()?.Trim()) != null)
+            {
+                if (SHADER_DEFINITION_END.Equals(line, StringComparison.InvariantCultureIgnoreCase))
+                {
+                    break;
+                }
+
+                if (!s_emptyLineRegex.IsMatch(line))
+                {
+                    Match groupMatch = s_groupRegex.Match(line);
+                    if (groupMatch.Success)
+                    {
+                        groups.Add(currentGroup = new Group(groupMatch.Groups[1].Value));
+                        continue;
+                    }
+
+                    Match shaderInfoMatch = s_shaderInfoRegex.Match(line);
+                    if (shaderInfoMatch.Success)
+                    {
+                        if (!Enum.TryParse(shaderInfoMatch.Groups[4].Value, out ShaderFlags flags))
+                        {
+                            throw new InvalidDataException(
+                                $"shader flags '{shaderInfoMatch.Groups[4].Value}' one or more flags are invalid or unsupported.");
+                        }
+                        currentGroup?.Add(
+                            new ShaderInfo(
+                                shaderInfoMatch.Groups[1].Value,
+                                shaderInfoMatch.Groups[2].Value,
+                                shaderInfoMatch.Groups[3].Value,
+                                flags));
+                    }
+                }
+            }
+
+            string shaderSource = sr.ReadToEnd();
+
+            return new Shader(
+                groups.Select(
+                    t =>
+                    {
+                        return (t.Name,
+                                t.ShaderInfos.Select(
+                                    s =>
+                                    {
+                                        using CompilationResult cr = ShaderBytecode.Compile(
+                                            shaderSource,
+                                            s.EntryPoint,
+                                            s.Profile,
+                                            s.Flags);
+                                        if (cr.HasErrors) { throw new InvalidDataException(cr.Message); }
+                                        return s.Type switch
+                                        {
+                                            "vs" => (Shader.Type.VertexShader,
+                                                     (ComObject)new VertexShader(graphicsDevice.Device, cr),
+                                                     ShaderSignature.GetInputSignature(cr),
+                                                     new ShaderReflection(cr)),
+                                            "ps" => (Shader.Type.PixelShader,
+                                                     (ComObject)new PixelShader(graphicsDevice.Device, cr),
+                                                     ShaderSignature.GetInputSignature(cr),
+                                                     new ShaderReflection(cr)),
+                                            "ds" => (Shader.Type.DomainShader,
+                                                     (ComObject)new DomainShader(graphicsDevice.Device, cr),
+                                                     ShaderSignature.GetInputSignature(cr),
+                                                     new ShaderReflection(cr)),
+                                            "gs" => (Shader.Type.GeometryShader,
+                                                     (ComObject)new GeometryShader(graphicsDevice.Device, cr),
+                                                     ShaderSignature.GetInputSignature(cr),
+                                                     new ShaderReflection(cr)),
+                                            "hs" => (Shader.Type.HullShader,
+                                                     (ComObject)new HullShader(graphicsDevice.Device, cr),
+                                                     ShaderSignature.GetInputSignature(cr),
+                                                     new ShaderReflection(cr)),
+                                            "cs" => (Shader.Type.ComputeShader,
+                                                     (ComObject)new ComputeShader(graphicsDevice.Device, cr),
+                                                     ShaderSignature.GetInputSignature(cr),
+                                                     new ShaderReflection(cr)),
+                                            _ => throw new InvalidDataException(
+                                                $"shader type '{s.Type}' doesn't exists or is unsupported.")
+                                        };
+                                    })
+                            );
+                    }));
+        }
+
         /// <summary>
         ///     A group. This class cannot be inherited.
         /// </summary>
@@ -137,102 +233,6 @@ namespace Exomia.Framework.Graphics.Shader
             {
                 return $"{Type} {EntryPoint} {Profile} {Flags}";
             }
-        }
-
-        public static Shader? FromStream(IGraphicsDevice graphicsDevice, Stream stream)
-        {
-            using StreamReader sr = new StreamReader(stream, Encoding.Default, true, 4096, true);
-            if (!SHADER_DEFINITION.Equals(sr.ReadLine()?.Trim(), StringComparison.InvariantCultureIgnoreCase))
-            {
-                return null;
-            }
-
-            IList<Group> groups = new List<Group>(1);
-
-            Group?  currentGroup = null;
-            string? line;
-            while ((line = sr.ReadLine()?.Trim()) != null)
-            {
-                if (SHADER_DEFINITION_END.Equals(line, StringComparison.InvariantCultureIgnoreCase))
-                {
-                    break;
-                }
-
-                if (!s_emptyLineRegex.IsMatch(line))
-                {
-                    Match groupMatch = s_groupRegex.Match(line);
-                    if (groupMatch.Success)
-                    {
-                        groups.Add(currentGroup = new Group(groupMatch.Groups[1].Value));
-                        continue;
-                    }
-
-                    Match shaderInfoMatch = s_shaderInfoRegex.Match(line);
-                    if (shaderInfoMatch.Success)
-                    {
-                        if (!Enum.TryParse(shaderInfoMatch.Groups[4].Value, out ShaderFlags flags))
-                        {
-                            throw new InvalidDataException(
-                                $"shader flags '{shaderInfoMatch.Groups[4].Value}' one or more flags are invalid or unsupported.");
-                        }
-                        currentGroup?.Add(
-                            new ShaderInfo(
-                                shaderInfoMatch.Groups[1].Value,
-                                shaderInfoMatch.Groups[2].Value,
-                                shaderInfoMatch.Groups[3].Value,
-                                flags));
-                    }
-                }
-            }
-
-            string shaderSource = sr.ReadToEnd();
-
-            return new Shader(
-                groups.Select(
-                    t =>
-                    {
-                        return (t.Name,
-                                t.ShaderInfos.Select(
-                                    s =>
-                                    {
-                                        using CompilationResult cr = ShaderBytecode.Compile(
-                                            shaderSource,
-                                            s.EntryPoint,
-                                            s.Profile,
-                                            s.Flags);
-                                        if (cr.HasErrors) { throw new InvalidDataException(cr.Message); }
-                                        return s.Type switch
-                                        {
-                                            "vs" => (Shader.Type.VertexShader,
-                                                     (ComObject)new VertexShader(graphicsDevice.Device, cr),
-                                                     ShaderSignature.GetInputSignature(cr),
-                                                     new ShaderReflection(cr)),
-                                            "ps" => (Shader.Type.PixelShader,
-                                                     (ComObject)new PixelShader(graphicsDevice.Device, cr),
-                                                     ShaderSignature.GetInputSignature(cr),
-                                                     new ShaderReflection(cr)),
-                                            "ds" => (Shader.Type.DomainShader,
-                                                     (ComObject)new DomainShader(graphicsDevice.Device, cr),
-                                                     ShaderSignature.GetInputSignature(cr),
-                                                     new ShaderReflection(cr)),
-                                            "gs" => (Shader.Type.GeometryShader,
-                                                     (ComObject)new GeometryShader(graphicsDevice.Device, cr),
-                                                     ShaderSignature.GetInputSignature(cr),
-                                                     new ShaderReflection(cr)),
-                                            "hs" => (Shader.Type.HullShader,
-                                                     (ComObject)new HullShader(graphicsDevice.Device, cr),
-                                                     ShaderSignature.GetInputSignature(cr),
-                                                     new ShaderReflection(cr)),
-                                            "cs" => (Shader.Type.ComputeShader,
-                                                     (ComObject)new ComputeShader(graphicsDevice.Device, cr),
-                                                     ShaderSignature.GetInputSignature(cr),
-                                                     new ShaderReflection(cr)),
-                                            _ => throw new InvalidDataException(
-                                                $"shader type '{s.Type}' doesn't exists or is unsupported.")
-                                        };
-                                    })
-                            );
-                    }));
         }
     }
 }
