+++ src/Exomia.Framework/Graphics/SpriteBatch.cs
@@ -15,7 +15,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Exomia.Framework.Graphics.Buffers;
@@ -31,7 +30,7 @@ namespace Exomia.Framework.Graphics
     /// <summary>
     ///     A sprite batch. This class cannot be inherited.
     /// </summary>
-    public sealed class SpriteBatch : IDisposable
+    public sealed partial class SpriteBatch : IDisposable
     {
         private const int MAX_BATCH_SIZE             = 1 << 13;
         private const int INITIAL_QUEUE_SIZE         = 1 << 7;
@@ -280,34 +279,6 @@ namespace Exomia.Framework.Graphics
             };
         }
 
-        internal struct SpriteInfo
-        {
-            public RectangleF    Source;
-            public RectangleF    Destination;
-            public Vector2       Origin;
-            public float         Rotation;
-            public float         Depth;
-            public SpriteEffects SpriteEffects;
-            public Color         Color;
-            public float         Opacity;
-        }
-
-        internal readonly struct TextureInfo
-        {
-            public readonly ShaderResourceView View;
-            public readonly int                Width;
-            public readonly int                Height;
-            public readonly long               Ptr64;
-
-            public TextureInfo(ShaderResourceView view, int width, int height)
-            {
-                View   = view;
-                Width  = width;
-                Height = height;
-                Ptr64  = view.NativePointer.ToInt64();
-            }
-        }
-
         /// <summary>
         ///     Draw batch per texture.
         /// </summary>
@@ -492,40 +463,6 @@ namespace Exomia.Framework.Graphics
             _context.InputAssembler.SetVertexBuffers(0, _vertexBuffer);
         }
 
-        [StructLayout(LayoutKind.Explicit, Size = VERTEX_STRIDE)]
-        private struct VertexPositionColorTexture
-        {
-            [FieldOffset(0)]
-            public float X;
-
-            [FieldOffset(4)]
-            public float Y;
-
-            [FieldOffset(8)]
-            public float Z;
-
-            [FieldOffset(12)]
-            public float W;
-
-            [FieldOffset(16)]
-            public float R;
-
-            [FieldOffset(20)]
-            public float G;
-
-            [FieldOffset(24)]
-            public float B;
-
-            [FieldOffset(28)]
-            public float A;
-
-            [FieldOffset(32)]
-            public float U;
-
-            [FieldOffset(36)]
-            public float V;
-        }
-
         /// <summary>
         ///     Updates the vertex from sprite information.
         /// </summary>
@@ -607,846 +544,67 @@ namespace Exomia.Framework.Graphics
             }
         }
 
-        #region Drawing
-
-        #region Defaults
-
-        /// <summary>
-        ///     Draw rectangle.
-        /// </summary>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="color">                The color. </param>
-        /// <param name="lineWidth">            Width of the line. </param>
-        /// <param name="rotation">             The rotation. </param>
-        /// <param name="opacity">              The opacity. </param>
-        /// <param name="layerDepth">           Depth of the layer. </param>
-        public void DrawRectangle(in RectangleF destinationRectangle,
-                                  in Color      color,
-                                  float         lineWidth,
-                                  float         rotation,
-                                  float         opacity,
-                                  float         layerDepth)
+        internal struct SpriteInfo
         {
-            DrawRectangle(destinationRectangle, color, lineWidth, rotation, s_vector2Zero, opacity, layerDepth);
+            public RectangleF    Source;
+            public RectangleF    Destination;
+            public Vector2       Origin;
+            public float         Rotation;
+            public float         Depth;
+            public SpriteEffects SpriteEffects;
+            public Color         Color;
+            public float         Opacity;
         }
 
-        /// <summary>
-        ///     Draw rectangle.
-        /// </summary>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="color">                The color. </param>
-        /// <param name="lineWidth">            Width of the line. </param>
-        /// <param name="rotation">             The rotation. </param>
-        /// <param name="origin">               The origin. </param>
-        /// <param name="opacity">              The opacity. </param>
-        /// <param name="layerDepth">           Depth of the layer. </param>
-        public void DrawRectangle(in RectangleF destinationRectangle,
-                                  in Color      color,
-                                  float         lineWidth,
-                                  float         rotation,
-                                  in Vector2    origin,
-                                  float         opacity,
-                                  float         layerDepth)
+        internal readonly struct TextureInfo
         {
-            Vector2[] vertex;
+            public readonly ShaderResourceView View;
+            public readonly int                Width;
+            public readonly int                Height;
+            public readonly long               Ptr64;
 
-            // ReSharper disable once CompareOfFloatsByEqualityOperator
-            if (rotation == 0.0f)
+            public TextureInfo(ShaderResourceView view, int width, int height)
             {
-                vertex = new[]
-                {
-                    destinationRectangle.TopLeft, destinationRectangle.TopRight, destinationRectangle.BottomRight,
-                    destinationRectangle.BottomLeft
-                };
-            }
-            else
-            {
-                vertex = new Vector2[4];
-
-                Vector2 o = origin;
-
-                // ReSharper disable once CompareOfFloatsByEqualityOperator
-                if (destinationRectangle.Width != 0f)
-                {
-                    o.X /= destinationRectangle.Width;
-                }
-
-                // ReSharper disable once CompareOfFloatsByEqualityOperator
-                if (destinationRectangle.Height != 0f)
-                {
-                    o.Y /= destinationRectangle.Height;
-                }
-
-                float cos = (float)Math.Cos(rotation);
-                float sin = (float)Math.Sin(rotation);
-                for (int j = 0; j < VERTICES_PER_SPRITE; j++)
-                {
-                    Vector2 corner = s_cornerOffsets[j];
-                    float   posX   = (corner.X - o.X) * destinationRectangle.Width;
-                    float   posY   = (corner.Y - o.Y) * destinationRectangle.Height;
-
-                    vertex[j] = new Vector2(
-                        (destinationRectangle.X + (posX * cos)) - (posY * sin),
-                        destinationRectangle.Y + (posX * sin) + (posY * cos));
-                }
-            }
-
-            DrawPolygon(vertex, color, lineWidth, opacity, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw fill rectangle.
-        /// </summary>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="color">                The color. </param>
-        /// <param name="layerDepth">           Depth of the layer. </param>
-        public void DrawFillRectangle(in RectangleF destinationRectangle, in Color color, float layerDepth)
-        {
-            DrawSprite(
-                _whiteTexture, destinationRectangle, false, s_nullRectangle,
-                color, 0.0f, s_vector2Zero, 1.0f, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw fill rectangle.
-        /// </summary>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="color">                The color. </param>
-        /// <param name="opacity">              The opacity. </param>
-        /// <param name="layerDepth">           Depth of the layer. </param>
-        public void DrawFillRectangle(in RectangleF destinationRectangle,
-                                      in Color      color,
-                                      float         opacity,
-                                      float         layerDepth)
-        {
-            DrawSprite(
-                _whiteTexture, destinationRectangle, false, s_nullRectangle,
-                color, 0.0f, s_vector2Zero, opacity, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw fill rectangle.
-        /// </summary>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="color">                The color. </param>
-        /// <param name="rotation">             The rotation. </param>
-        /// <param name="origin">               The origin. </param>
-        /// <param name="opacity">              The opacity. </param>
-        /// <param name="layerDepth">           Depth of the layer. </param>
-        public void DrawFillRectangle(in RectangleF destinationRectangle,
-                                      in Color      color,
-                                      float         rotation,
-                                      in Vector2    origin,
-                                      float         opacity,
-                                      float         layerDepth)
-        {
-            DrawSprite(
-                _whiteTexture, destinationRectangle, false, s_nullRectangle,
-                color, rotation, origin, opacity, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw line.
-        /// </summary>
-        /// <param name="point1">     The first point. </param>
-        /// <param name="point2">     The second point. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="lineWidth">  Width of the line. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawLine(in Vector2 point1,
-                             in Vector2 point2,
-                             in Color   color,
-                             float      lineWidth,
-                             float      opacity,
-                             float      layerDepth)
-        {
-            DrawLine(point1, point2, color, lineWidth, opacity, 1.0f, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw line.
-        /// </summary>
-        /// <param name="point1">       The first point. </param>
-        /// <param name="point2">       The second point. </param>
-        /// <param name="color">        The color. </param>
-        /// <param name="lineWidth">    Width of the line. </param>
-        /// <param name="opacity">      The opacity. </param>
-        /// <param name="lengthFactor"> The length factor. </param>
-        /// <param name="layerDepth">   Depth of the layer. </param>
-        public void DrawLine(in Vector2 point1,
-                             in Vector2 point2,
-                             in Color   color,
-                             float      lineWidth,
-                             float      opacity,
-                             float      lengthFactor,
-                             float      layerDepth)
-        {
-            DrawSprite(
-                _whiteTexture, new RectangleF(
-                    point1.X, point1.Y, Vector2.Distance(point1, point2) * lengthFactor, lineWidth), false,
-                s_nullRectangle, color, (float)Math.Atan2(point2.Y - point1.Y, point2.X - point1.X),
-                s_vector2Zero, opacity, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw polygon.
-        /// </summary>
-        /// <param name="vertex">     The vertex. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="lineWidth">  Width of the line. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawPolygon(Vector2[] vertex, in Color color, float lineWidth, float opacity, float layerDepth)
-        {
-            if (vertex.Length > 1)
-            {
-                int l = vertex.Length - 1;
-                for (int i = 0; i < l; i++)
-                {
-                    DrawLine(vertex[i], vertex[i + 1], color, lineWidth, opacity, layerDepth);
-                }
-                DrawLine(vertex[l], vertex[0], color, lineWidth, opacity, layerDepth);
+                View   = view;
+                Width  = width;
+                Height = height;
+                Ptr64  = view.NativePointer.ToInt64();
             }
         }
 
-        /// <summary>
-        ///     Draw circle.
-        /// </summary>
-        /// <param name="center">     The center. </param>
-        /// <param name="radius">     The radius. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="lineWidth">  Width of the line. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="segments">   The segments. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawCircle(in Vector2 center,
-                               float      radius,
-                               in Color   color,
-                               float      lineWidth,
-                               float      opacity,
-                               int        segments,
-                               float      layerDepth)
+        [StructLayout(LayoutKind.Explicit, Size = VERTEX_STRIDE)]
+        private struct VertexPositionColorTexture
         {
-            DrawCircle(center, radius, 0, MathUtil.TwoPi, color, lineWidth, opacity, segments, layerDepth);
+            [FieldOffset(0)]
+            public float X;
+
+            [FieldOffset(4)]
+            public float Y;
+
+            [FieldOffset(8)]
+            public float Z;
+
+            [FieldOffset(12)]
+            public float W;
+
+            [FieldOffset(16)]
+            public float R;
+
+            [FieldOffset(20)]
+            public float G;
+
+            [FieldOffset(24)]
+            public float B;
+
+            [FieldOffset(28)]
+            public float A;
+
+            [FieldOffset(32)]
+            public float U;
+
+            [FieldOffset(36)]
+            public float V;
         }
-
-        /// <summary>
-        ///     Draw circle.
-        /// </summary>
-        /// <param name="center">     The center. </param>
-        /// <param name="radius">     The radius. </param>
-        /// <param name="start">      The start. </param>
-        /// <param name="end">        The end. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="lineWidth">  Width of the line. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="segments">   The segments. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawCircle(in Vector2 center,
-                               float      radius,
-                               float      start,
-                               float      end,
-                               in Color   color,
-                               float      lineWidth,
-                               float      opacity,
-                               int        segments,
-                               float      layerDepth)
-        {
-            Vector2[] vertex = new Vector2[segments];
-
-            float increment = (end - start) / segments;
-            float theta     = start;
-
-            for (int i = 0; i < segments; i++)
-            {
-                vertex[i].X =  center.X + (radius * (float)Math.Cos(theta));
-                vertex[i].Y =  center.Y + (radius * (float)Math.Sin(theta));
-                theta       += increment;
-            }
-
-            DrawPolygon(vertex, color, lineWidth, opacity, layerDepth);
-        }
-
-        #endregion
-
-        #region Texture
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">  The texture. </param>
-        /// <param name="position"> The position. </param>
-        /// <param name="color">    The color. </param>
-        public void Draw(Texture texture, in Vector2 position, in Color color)
-        {
-            DrawSprite(
-                texture, new RectangleF(position.X, position.Y, 1f, 1f), true, s_nullRectangle,
-                color, 0f, s_vector2Zero, 1.0f, SpriteEffects.None, 0f);
-        }
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">              The texture. </param>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="color">                The color. </param>
-        public void Draw(Texture texture, in RectangleF destinationRectangle, in Color color)
-        {
-            DrawSprite(
-                texture, destinationRectangle, false, s_nullRectangle,
-                color, 0f, s_vector2Zero, 1.0f, SpriteEffects.None, 0f);
-        }
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">         The texture. </param>
-        /// <param name="position">        The position. </param>
-        /// <param name="sourceRectangle"> Source rectangle. </param>
-        /// <param name="color">           The color. </param>
-        public void Draw(Texture texture, in Vector2 position, in Rectangle? sourceRectangle, in Color color)
-        {
-            DrawSprite(
-                texture, new RectangleF(position.X, position.Y, 1f, 1f), true, sourceRectangle,
-                color, 0f, s_vector2Zero, 1.0f, SpriteEffects.None, 0f);
-        }
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">              The texture. </param>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="sourceRectangle">      Source rectangle. </param>
-        /// <param name="color">                The color. </param>
-        public void Draw(Texture       texture,
-                         in RectangleF destinationRectangle,
-                         in Rectangle? sourceRectangle,
-                         in Color      color)
-        {
-            DrawSprite(
-                texture, destinationRectangle, false, sourceRectangle,
-                color, 0f, s_vector2Zero, 1.0f, SpriteEffects.None, 0f);
-        }
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">              The texture. </param>
-        /// <param name="destinationRectangle"> Destination rectangle. </param>
-        /// <param name="sourceRectangle">      Source rectangle. </param>
-        /// <param name="color">                The color. </param>
-        /// <param name="rotation">             The rotation. </param>
-        /// <param name="origin">               The origin. </param>
-        /// <param name="opacity">              The opacity. </param>
-        /// <param name="effects">              The effects. </param>
-        /// <param name="layerDepth">           Depth of the layer. </param>
-        public void Draw(Texture       texture,
-                         in RectangleF destinationRectangle,
-                         in Rectangle? sourceRectangle,
-                         in Color      color,
-                         float         rotation,
-                         in Vector2    origin,
-                         float         opacity,
-                         SpriteEffects effects,
-                         float         layerDepth)
-        {
-            DrawSprite(
-                texture, destinationRectangle, false, sourceRectangle,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">         The texture. </param>
-        /// <param name="position">        The position. </param>
-        /// <param name="sourceRectangle"> Source rectangle. </param>
-        /// <param name="color">           The color. </param>
-        /// <param name="rotation">        The rotation. </param>
-        /// <param name="origin">          The origin. </param>
-        /// <param name="scale">           The scale. </param>
-        /// <param name="opacity">         The opacity. </param>
-        /// <param name="effects">         The effects. </param>
-        /// <param name="layerDepth">      Depth of the layer. </param>
-        public void Draw(Texture       texture,
-                         in Vector2    position,
-                         in Rectangle? sourceRectangle,
-                         in Color      color,
-                         float         rotation,
-                         in Vector2    origin,
-                         float         scale,
-                         float         opacity,
-                         SpriteEffects effects,
-                         float         layerDepth)
-        {
-            DrawSprite(
-                texture, new RectangleF(position.X, position.Y, scale, scale), true, sourceRectangle,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draws.
-        /// </summary>
-        /// <param name="texture">         The texture. </param>
-        /// <param name="position">        The position. </param>
-        /// <param name="sourceRectangle"> Source rectangle. </param>
-        /// <param name="color">           The color. </param>
-        /// <param name="rotation">        The rotation. </param>
-        /// <param name="origin">          The origin. </param>
-        /// <param name="scale">           The scale. </param>
-        /// <param name="opacity">         The opacity. </param>
-        /// <param name="effects">         The effects. </param>
-        /// <param name="layerDepth">      Depth of the layer. </param>
-        public void Draw(Texture       texture,
-                         in Vector2    position,
-                         in Rectangle? sourceRectangle,
-                         in Color      color,
-                         float         rotation,
-                         in Vector2    origin,
-                         in Vector2    scale,
-                         float         opacity,
-                         SpriteEffects effects,
-                         float         layerDepth)
-        {
-            DrawSprite(
-                texture, new RectangleF(position.X, position.Y, scale.X, scale.Y), true, sourceRectangle,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw sprite.
-        /// </summary>
-        /// <exception cref="ArgumentNullException">
-        ///     Thrown when one or more required arguments are
-        ///     null.
-        /// </exception>
-        /// <exception cref="InvalidOperationException"> Thrown when the requested operation is invalid. </exception>
-        /// <param name="texture">          The texture. </param>
-        /// <param name="destination">      Destination for the. </param>
-        /// <param name="scaleDestination"> True to scale destination. </param>
-        /// <param name="sourceRectangle">  Source rectangle. </param>
-        /// <param name="color">            The color. </param>
-        /// <param name="rotation">         The rotation. </param>
-        /// <param name="origin">           The origin. </param>
-        /// <param name="opacity">          The opacity. </param>
-        /// <param name="effects">          The effects. </param>
-        /// <param name="depth">            The depth. </param>
-        private unsafe void DrawSprite(Texture       texture,
-                                       in RectangleF destination,
-                                       bool          scaleDestination,
-                                       in Rectangle? sourceRectangle,
-                                       in Color      color,
-                                       float         rotation,
-                                       in Vector2    origin,
-                                       float         opacity,
-                                       SpriteEffects effects,
-                                       float         depth)
-        {
-            if (!_isBeginCalled)
-            {
-                throw new InvalidOperationException("Begin must be called before draw");
-            }
-
-            if (texture.TexturePointer == IntPtr.Zero)
-            {
-                throw new ArgumentNullException(nameof(texture));
-            }
-
-            if (_spriteQueueCount >= _spriteQueue.Length)
-            {
-                bool lockTaken = false;
-                try
-                {
-                    _spinLock.Enter(ref lockTaken);
-
-                    int size = _spriteQueue.Length * 2;
-                    _sortIndices   = new int[size];
-                    _sortedSprites = new SpriteInfo[size];
-                    Array.Resize(ref _spriteQueue, size);
-                    Array.Resize(ref _spriteTextures, size);
-                }
-                finally
-                {
-                    if (lockTaken)
-                    {
-                        _spinLock.Exit(false);
-                    }
-                }
-            }
-
-            if (!_textureInfos.TryGetValue(texture.TexturePointer, out TextureInfo textureInfo))
-            {
-                bool lockTaken = false;
-                try
-                {
-                    _spinLock.Enter(ref lockTaken);
-                    if (!_textureInfos.TryGetValue(texture.TexturePointer, out textureInfo))
-                    {
-                        textureInfo = new TextureInfo(texture.TextureView, texture.Width, texture.Height);
-                        _textureInfos.Add(texture.TexturePointer, textureInfo);
-                    }
-                }
-                finally
-                {
-                    if (lockTaken)
-                    {
-                        _spinLock.Exit(false);
-                    }
-                }
-            }
-
-            int spriteQueueCount = Interlocked.Increment(ref _spriteQueueCount) - 1;
-            fixed (SpriteInfo* spriteInfo = &_spriteQueue[spriteQueueCount])
-            {
-                float width;
-                float height;
-                if (sourceRectangle.HasValue)
-                {
-                    Rectangle rectangle = sourceRectangle.Value;
-                    spriteInfo->Source.X = rectangle.X;
-                    spriteInfo->Source.Y = rectangle.Y;
-                    width                = rectangle.Width;
-                    height               = rectangle.Height;
-                }
-                else
-                {
-                    spriteInfo->Source.X = 0;
-                    spriteInfo->Source.Y = 0;
-                    width                = texture.Width;
-                    height               = texture.Height;
-                }
-
-                spriteInfo->Source.Width  = width;
-                spriteInfo->Source.Height = height;
-
-                spriteInfo->Destination.X = destination.X;
-                spriteInfo->Destination.Y = destination.Y;
-
-                if (scaleDestination)
-                {
-                    spriteInfo->Destination.Width  = destination.Width * width;
-                    spriteInfo->Destination.Height = destination.Height * height;
-                }
-                else
-                {
-                    spriteInfo->Destination.Width  = destination.Width;
-                    spriteInfo->Destination.Height = destination.Height;
-                }
-
-                if (spriteInfo->Destination.Width < 0)
-                {
-                    spriteInfo->Destination.X     += spriteInfo->Destination.Width;
-                    spriteInfo->Destination.Width =  -spriteInfo->Destination.Width;
-                }
-
-                if (spriteInfo->Destination.Height < 0)
-                {
-                    spriteInfo->Destination.Y      += spriteInfo->Destination.Height;
-                    spriteInfo->Destination.Height =  -spriteInfo->Destination.Height;
-                }
-
-                spriteInfo->Origin        = origin;
-                spriteInfo->Rotation      = rotation;
-                spriteInfo->Depth         = depth;
-                spriteInfo->SpriteEffects = effects;
-                spriteInfo->Color         = color;
-                spriteInfo->Opacity       = opacity;
-            }
-
-            _spriteTextures[spriteQueueCount] = textureInfo;
-        }
-
-        #endregion
-
-        #region SpiteFont
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont font, string text, in Vector2 position, in Color color, float layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, position,
-                color, 0f, Vector2.Zero, 1.0f, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont font,
-                             string     text,
-                             in Vector2 position,
-                             in Color   color,
-                             float      rotation,
-                             float      layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, position,
-                color, rotation, Vector2.Zero, 1.0f, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="origin">     The origin. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="effects">    The effects. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             string        text,
-                             in Vector2    position,
-                             in Color      color,
-                             float         rotation,
-                             in Vector2    origin,
-                             float         opacity,
-                             SpriteEffects effects,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, position,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="start">      The start. </param>
-        /// <param name="end">        The end. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="origin">     The origin. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="effects">    The effects. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             string        text,
-                             int           start,
-                             int           end,
-                             in Vector2    position,
-                             in Color      color,
-                             float         rotation,
-                             in Vector2    origin,
-                             float         opacity,
-                             SpriteEffects effects,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, start, end, position,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="start">      The start. </param>
-        /// <param name="end">        The end. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="dimension">  The dimension. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="origin">     The origin. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="effects">    The effects. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             string        text,
-                             int           start,
-                             int           end,
-                             in Vector2    position,
-                             in Size2F     dimension,
-                             in Color      color,
-                             float         rotation,
-                             in Vector2    origin,
-                             float         opacity,
-                             SpriteEffects effects,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, start, end, position, dimension,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont font, StringBuilder text, in Vector2 position, in Color color, float layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, position,
-                color, 0f, Vector2.Zero, 1.0f, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             StringBuilder text,
-                             in Vector2    position,
-                             in Color      color,
-                             float         rotation,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, position,
-                color, rotation, Vector2.Zero, 1.0f, SpriteEffects.None, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="origin">     The origin. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="effects">    The effects. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             StringBuilder text,
-                             in Vector2    position,
-                             in Color      color,
-                             float         rotation,
-                             in Vector2    origin,
-                             float         opacity,
-                             SpriteEffects effects,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, position,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="start">      The start. </param>
-        /// <param name="end">        The end. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="origin">     The origin. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="effects">    The effects. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             StringBuilder text,
-                             int           start,
-                             int           end,
-                             in Vector2    position,
-                             in Color      color,
-                             float         rotation,
-                             in Vector2    origin,
-                             float         opacity,
-                             SpriteEffects effects,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, start, end, position,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text.
-        /// </summary>
-        /// <param name="font">       The font. </param>
-        /// <param name="text">       The text. </param>
-        /// <param name="start">      The start. </param>
-        /// <param name="end">        The end. </param>
-        /// <param name="position">   The position. </param>
-        /// <param name="dimension">  The dimension. </param>
-        /// <param name="color">      The color. </param>
-        /// <param name="rotation">   The rotation. </param>
-        /// <param name="origin">     The origin. </param>
-        /// <param name="opacity">    The opacity. </param>
-        /// <param name="effects">    The effects. </param>
-        /// <param name="layerDepth"> Depth of the layer. </param>
-        public void DrawText(SpriteFont    font,
-                             StringBuilder text,
-                             int           start,
-                             int           end,
-                             in Vector2    position,
-                             in Size2F     dimension,
-                             in Color      color,
-                             float         rotation,
-                             in Vector2    origin,
-                             float         opacity,
-                             SpriteEffects effects,
-                             float         layerDepth)
-        {
-            font.Draw(
-                DrawTextInternal, text, start, end, position, dimension,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        /// <summary>
-        ///     Draw text internal.
-        /// </summary>
-        /// <param name="texture">         The texture. </param>
-        /// <param name="position">        The position. </param>
-        /// <param name="sourceRectangle"> Source rectangle. </param>
-        /// <param name="color">           The color. </param>
-        /// <param name="rotation">        The rotation. </param>
-        /// <param name="origin">          The origin. </param>
-        /// <param name="scale">           The scale. </param>
-        /// <param name="opacity">         The opacity. </param>
-        /// <param name="effects">         The effects. </param>
-        /// <param name="layerDepth">      Depth of the layer. </param>
-        internal void DrawTextInternal(Texture       texture,
-                                       in Vector2    position,
-                                       in Rectangle? sourceRectangle,
-                                       in Color      color,
-                                       float         rotation,
-                                       in Vector2    origin,
-                                       float         scale,
-                                       float         opacity,
-                                       SpriteEffects effects,
-                                       float         layerDepth)
-        {
-            DrawSprite(
-                texture, new RectangleF(position.X, position.Y, scale, scale), true, sourceRectangle,
-                color, rotation, origin, opacity, effects, layerDepth);
-        }
-
-        #endregion
-
-        #endregion
 
         #region IDisposable Support
